requirements.txt
gymnasium  # tested with 0.28.1
matplotlib  # tested with 3.5.3
numpy  # tested with 1.24.4
Pillow  # tested with 9.5.0